requirements.txt
requests>=2.31.0
cryptography>=41.0.0
websocket-client>=1.7.0
flask>=3.0.0
